fix: component search box
