+++ src/hooks/useCollist.tsx
@@ -1,17 +1,9 @@
-import React from 'react';
-import { Link } from 'react-router-dom'; // React Router中的Link
+
 import type { ColumnsType } from 'antd/es/table';
 import { useNavigate } from 'react-router-dom';
-import { useEffect, useState, useRef } from 'react';
-import { Input, Button, Space } from 'antd';
-import { SearchOutlined } from '@ant-design/icons';
-import Highlighter from 'react-highlight-words';
-
-type FilterConfirmProps = { closeDropdown: boolean };
-
-
-type DataIndex = keyof DataType | number;
-interface DataType {
+import { useEffect, useState } from 'react';
+import getColumnSearchProps from '../components/getColumnSearchProps';
+type DataType = {
   key: React.Key;
   name: string;
   houseid: string;
@@ -19,9 +11,15 @@ interface DataType {
   amount: string;
   id: string;
 }
-
+type DataTypes = {
+  title: string;
+  dataIndex: string;
+  key: string;
+  width?: string;
+  type: string;
+  amount: string;
+}
 const useCollist = () => {
-
   const [data, setData] = useState<DataType[]>([
     {
       key: '1',
@@ -57,108 +55,8 @@ const useCollist = () => {
     },
   ]);
 
-  const [searchText, setSearchText] = useState('');
-  const [searchedColumn, setSearchedColumn] = useState('');
-  const searchInput = useRef<InputRef>(null);
 
-  const handleSearch = (
-    selectedKeys: string[],
-    confirm: (param?: FilterConfirmProps) => void,
-    dataIndex: DataIndex,
-  ) => {
-    confirm();
-    setSearchText(selectedKeys[0]);
-    setSearchedColumn(dataIndex);
-  };
-
-  const handleReset = (clearFilters: () => void) => {
-    clearFilters();
-    setSearchText('');
-  };
-
-  const getColumnSearchProps = (dataIndex: DataIndex): ColumnsType<DataType> => ({
-    filterDropdown: ({ setSelectedKeys, selectedKeys, confirm, clearFilters, close }: {
-      setSelectedKeys: (keys: string[]) => void;
-      selectedKeys: string[];
-      confirm: () => void;
-      clearFilters: () => void;
-      close: () => void;
-    }) => (
-      <div style={{ padding: 8 }} onKeyDown={(e) => e.stopPropagation()}>
-        <Input
-          ref={searchInput}
-          placeholder={`Search ${dataIndex}`}
-          value={selectedKeys[0]}
-          onChange={(e) => setSelectedKeys(e.target.value ? [e.target.value] : [])}
-          onPressEnter={() => handleSearch(selectedKeys as string[], confirm, dataIndex)}
-          style={{ marginBottom: 8, display: 'block' }}
-        />
-        <Space>
-          <Button
-            type="primary"
-            onClick={() => handleSearch(selectedKeys as string[], confirm, dataIndex)}
-            icon={<SearchOutlined />}
-            size="small"
-            style={{ width: 90 }}
-          >
-            搜尋
-          </Button>
-          <Button
-            onClick={() => clearFilters && handleReset(clearFilters)}
-            size="small"
-            style={{ width: 90 }}
-          >
-            重設
-          </Button>
-          {/* <Button
-          type="link"
-          size="small"
-          onClick={() => {
-            confirm({ closeDropdown: false });
-            setSearchText((selectedKeys as string[])[0]);
-            setSearchedColumn(dataIndex);
-          }}
-        >
-          Filter
-        </Button> */}
-          <Button
-            type="link"
-            size="small"
-            onClick={() => {
-              close();
-            }}
-          >
-            關閉
-          </Button>
-        </Space>
-      </div>
-    ),
-    filterIcon: (filtered: boolean) => (
-      <SearchOutlined style={{ color: filtered ? '#1677ff' : undefined }} />
-    ),
-    onFilter: (value: any, record: any) =>
-      record[dataIndex]
-        .toString()
-        .toLowerCase()
-        .includes((value as string).toLowerCase()),
-    onFilterDropdownOpenChange: (visible: any) => {
-      if (visible) {
-        setTimeout(() => searchInput.current?.select(), 100);
-      }
-    },
-    render: (text: any) =>
-      searchedColumn === dataIndex ? (
-        <Highlighter
-          highlightStyle={{ backgroundColor: '#ffc069', padding: 0 }}
-          searchWords={[searchText]}
-          autoEscape
-          textToHighlight={text ? text.toString() : ''}
-        />
-      ) : (
-        text
-      ),
-  });
-  const columns: ColumnsType<DataType> = [
+  const columns: ColumnsType<DataTypes> = [
     {
       title: '編號',
       dataIndex: 'id',
@@ -196,7 +94,6 @@ const useCollist = () => {
 
     }
   ];
-
   const getListdata = async () => {
     const response = await fetch('http://localhost:5173/api/colelist');
     if (!response.ok) {
